Feat: Set up team main page
