+++ src/app/teams/[teamId]/panel/TeamDnD.tsx
@@ -3,8 +3,8 @@ import useAxiosWithAuth from '@/api/config'
 import 'react-grid-layout/css/styles.css'
 import { ITeamDnDLayout, SizeType, WidgetType } from '@/types/ITeamDnDLayout'
 import { useEffect, useState } from 'react'
-import WidgetList from '@/app/teams/[teamId]/[id]/panel/widgets/WidgetList'
-import WidgetsRender from '@/app/teams/[teamId]/[id]/panel/widgets/WidgetsRender'
+import WidgetList from '@/app/teams/[teamId]/panel/widgets/WidgetList'
+import WidgetsRender from '@/app/teams/[teamId]/panel/widgets/WidgetsRender'
 import ReactGridLayout from 'react-grid-layout'
 import useSWRMutation from 'swr/mutation'
