+++ src/app/teams/[teamId]/panel/TeamInfoComponent.tsx
