+++ src/app/teams/[teamId]/panel/widgets/TmpNoticeWidget.tsx
@@ -1,6 +1,6 @@
 import { TextField, Typography } from '@mui/material'
 import { SizeType } from '@/types/ITeamDnDLayout'
-import WidgetCard from '@/app/teams/[teamId]/[id]/panel/widgets/WidgetCard'
+import WidgetCard from '@/app/teams/[teamId]/panel/widgets/WidgetCard'
 
 /* 임시 위젯 */
 const TmpNoticeWidget = ({ data, size }: { data: any; size: SizeType }) => {
