+++ src/app/teams/[teamId]/panel/widgets/WidgetCard.tsx
