+++ src/app/teams/[teamId]/style/CalendarLarge.scss
