+++ src/app/teams/[teamId]/style/CalendarMini.scss
